makino-ai-companion/requirements.txt
# Core - Claude API & Web Framework
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0.1
jinja2>=3.1.0

# Knowledge Pipeline - ドキュメント処理
PyMuPDF>=1.24.0        # PDF テキスト+画像抽出
python-docx>=1.1.0     # Word文書テキスト+画像抽出
Pillow>=10.0.0         # 画像処理

# Knowledge Pipeline - 音声/動画処理（動画・音声ファイルを扱う場合）
# openai-whisper>=20231117  # 音声→テキスト文字起こし（要ffmpeg）
# ffmpeg-python>=0.2.0     # 動画→音声抽出

# RAG - ベクトル検索
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0

# Auth - Google OAuth2 + JWT
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.0

# Development
pytest>=7.0.0
pytest-cov>=4.1.0
httpx>=0.27.0  # FastAPIテスト用
